+++ lib/pages/counter/counter_page.dart
@@ -1,20 +1,27 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:nft/pages/counter/counter_provider.dart';
+import 'package:nft/widgets/route_active_mixin.dart';
 import 'package:nft/widgets/screen_widget.dart';
 import 'package:provider/provider.dart';
 
-class CounterPage extends StatelessWidget {
+class CounterPage extends StatefulWidget {
   const CounterPage({Key key, this.argument}) : super(key: key);
 
   final String argument;
 
   @override
+  _CounterPageState createState() => _CounterPageState();
+}
+
+class _CounterPageState extends State<CounterPage>
+    with RouteActiveMixin<CounterPage> {
+  @override
   Widget build(BuildContext context) {
     return ScreenWidget(
       body: Column(children: <Widget>[
         AppBar(),
-        ScreenHeader(title: argument ?? ''),
+        ScreenHeader(title: widget.argument ?? ''),
         const Expanded(
           child: ScreenBody(),
         ),